pyproject.toml
[tool.poetry]
name = "spkc-ehr-schema"
version = "1.0.15"
description = "Engine for EHR schema manipulations"
authors = ["SpeakCare Team"]
readme = "README.md"
packages = [{include = "schema_engine", from = "src"}, {include = "pcc_schema", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
jsonschema = "4.25.1"
jsonschema-specifications = "2025.9.1"
boto3 = "1.35.0"



[tool.poetry.group.dev.dependencies]
pytest = "8.4.2"
pytest-cov = "7.0.0"
pytest-mock = "3.15.0"
black = "25.1.0"
flake8 = "7.3.0"
mypy = "1.17.1"
openai = "^2.6.1"
python-dotenv = "^1.2.1"
google-genai = "^1.60.0"



[[tool.poetry.source]]
name = "codeartifact"
url = "https://speakcare-182399723275.d.codeartifact.us-east-1.amazonaws.com/pypi/python-packages/simple/"
priority = "supplemental"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 88
target-version = ['py313']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["*_test.py", "test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
